% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllJointAnomalies.R
\name{findAllJointAnomalies}
\alias{findAllJointAnomalies}
\title{Joint anomaly discovery function for my scolarship}
\usage{
findAllJointAnomalies(file, config)
}
\arguments{
\item{file}{The absolute path of the dataset.}

\item{config}{The absolute path of the config file}

\item{var1}{The name of the main variable}
}
\description{
This function allows you find common anomalies between all variables of the dataset
}
\examples{
findJointAnomalies()
}
\keyword{DBSCAN}
\keyword{anomalies}
\keyword{anomaly}

